Move schema annotation to io.sigs.seals
